==> analyses/01-gbif-wrangling.Rmd <==
---
title: "Investigate GBIF records"
author: "Natalie Cooper"
date: "7 March 2019"
output: html_document
---

```{r, echo = FALSE, message = FALSE, warning = FALSE}
# Load libraries
library(tidyverse)
library(viridis)
library(knitr)
#library(sf)
library(maptools)
library(rgdal)
library(sp)
library(raster)
library(rworldmap)
library(rgeos)
library(spatstat)
library(ggmap)
library(here)

# Short function to get decades
floor_decade <- function(x){
  if(!is.na(x)){
    x - x %% 10
  }else{NA_character_}
}

# Convex hulls function...
make.hull <- function(data){
  hull <- spatstat::convexhull.xy(data$decimallongitude, data$decimallatitude)
  data.frame(x = hull$bdry[[1]]$x, y = hull$bdry[[1]]$y)
}

# Convert SpatialPolygonsDataFrame to dataframe for plotting
convert.spdf <- function(map) {
  map@data$id <- rownames(map@data)
  map.points <- fortify(map, region = "id")
  map.df <- dplyr::full_join(map.points, map@data, by = "id")
  return(map.df)
}
```

First we need to wrangle the GBIF data so only specimens with lat longs and full species names are included, we also add a "clade" variable to deal with fishes being in different classes.
```{r, echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
# Read in the GBIF data
gbif <- read_csv(here("data/all-gbif.csv"))

# Wrangle the data so it is useable
gbif2 <- 
  gbif %>%
  
  # Check that all included records are specimens not observations
  filter(basisofrecord == "PRESERVED_SPECIMEN") %>%
  
  # Create a new column for specimen ID number
  unite(col = specID, `institutioncode`, `catalognumber`, 
        sep = "_", remove = FALSE) %>%
  
  # Remove post 1900 records and those with no year
  filter(year < 1901 & !is.na(year))  %>%

  # Add decade variable (function above)
  # This maps to character to deal with NAs so needs coercing back to numeric
  mutate(decade = map_chr(year, floor_decade)) %>%
  mutate(decade = as.numeric(decade)) %>%

  # Add clade column, remove sea squirts and make a "fishes" "clade"
  filter(class != "Ascidiacea" & class != "Thaliacea" & !is.na(class)) %>%
  mutate(clade = class) %>%
  mutate(clade = if_else(clade == "Amphibia", "Amphibians", clade)) %>%
  mutate(clade = if_else(clade == "Reptilia", "Reptiles", clade)) %>%
  mutate(clade = if_else(clade == "Aves", "Birds", clade)) %>%
  mutate(clade = if_else(clade == "Mammalia", "Mammals", clade)) %>%
  mutate(clade = if_else(clade == "Actinopterygii", "Fishes", clade)) %>%
  mutate(clade = if_else(clade == "Cephalaspidomorphi", "Fishes", clade)) %>%
  mutate(clade = if_else(clade == "Elasmobranchii", "Fishes", clade)) %>%
  mutate(clade = if_else(clade == "Leptocardii", "Fishes", clade)) %>%

  # Remove records with no lat/long
  filter(!is.na(decimallatitude) & !is.na(decimallongitude)) %>%
  
  # TAXONOMY (note there is a second set of checks later)
  # 1. Remove entries where we only have the Genus
  # 2. Create binomial column
  separate(species, c("Genus", "Species"), sep = " ", 
           extra = "drop", fill = "right") %>%
    filter(!is.na(Species) & Species != "sp." & Species != "sp") %>%
    unite("binomial", Genus, Species, sep = " ") %>%

  # Select just the columns of interest
  dplyr::select(specID, binomial, clade, class, order, family, genus,
           year, decade, decimallongitude, decimallatitude) %>%
  
  # Remove duplicates
  distinct()

# Remove gbif
rm(gbif)

# Write out tidy gbif data
write_csv(gbif2, path = here("data/tidy-gbif.csv"))
```

## What data are available in GBIF?

Specimens per clade
```{r, echo = FALSE}
sum_clade <- gbif2 %>%
     group_by(clade) %>%
     distinct() %>%
     summarise(no.specimens = n()) 

kable(sum_clade)
```

Total specimens
```{r}
sum(sum_clade$no.specimens)
```

Species per clade
```{r, echo = FALSE}
sum_clade <- gbif2 %>%
     group_by(clade) %>%
     distinct() %>%
     summarise(no.species = length(unique(binomial))) 

kable(sum_clade)
```

Total species
```{r}
sum(sum_clade$no.species)
```


What about with more than 3 specimens?
```{r, echo = FALSE}
sum_clade <- gbif2 %>%
     group_by(clade) %>%
     summarise(n = n()) %>%
     filter(n > 3)

kable(sum_clade)
```

What about with more than 3 specimens per species with distinct locality data?
```{r, echo = FALSE}
sum_clade3 <- gbif2 %>%
     distinct(clade, binomial, decimallongitude, decimallatitude) %>%
     group_by(clade, binomial) %>%
     summarise(n = n()) %>%
     filter(n > 3) %>%
     group_by(clade) %>%
     summarise(n = n())

kable(sum_clade3)
```

What about with 10 or more specimens per species with distinct locality data?
```{r, echo = FALSE}
sum_clade10 <- gbif2 %>%
     distinct(clade, binomial, decimallongitude, decimallatitude) %>%
     group_by(clade, binomial) %>%
     summarise(n = n()) %>%
     filter(n >= 10) %>%
     group_by(clade) %>%
    summarise(n = n())

kable(sum_clade10)
```


## Maps of specimen distributions

All specimens where there are >= 3 unique localities
```{r, echo = FALSE}
# First find only those species with >= 3 specimens
specimen3 <- 
  gbif2 %>%
  distinct(binomial, decimallongitude, decimallatitude) %>%
  group_by(binomial) %>%
  summarise(n = n()) %>%
  filter(n >= 3)

# Then select these specimens from the gbif data
gbif3 <- gbif2[gbif2$binomial %in% specimen3$binomial, ]
```

```{r, echo = FALSE, warning = FALSE, message = FALSE}
# Select world map
world <- map_data("world")
```

```{r, echo = FALSE}
ggplot(data = world, aes(x = long, y = lat, group = group)) + 
  geom_polygon(fill = NA, col = "black", size = 0.5) +
  geom_point(data = gbif3, aes(x = decimallongitude, y = decimallatitude, 
                                  group = NA, color = binomial), alpha = 0.9) +
  facet_wrap( ~ clade) +
  theme_void() +
  theme(legend.position = "none")
```

How are these spread across time?

Just for birds as they swamp the data
```{r, echo = FALSE}
ggplot(data = world, aes(x = long, y = lat, group = group)) + 
  geom_polygon(fill = NA, col = "black", size = 0.5) +
  geom_point(data = gbif3[gbif3$clade == "Birds", ], aes(x = decimallongitude, y = decimallatitude, 
                                  group = NA, color = binomial), alpha = 0.5) +
  facet_wrap( ~ decade) +
  theme_void() +
  theme(legend.position = "none")
```


And for all the non birds
```{r, echo = FALSE}
ggplot(data = world, aes(x = long, y = lat, group = group)) + 
  geom_polygon(fill = NA, col = "black", size = 0.5) +
  geom_point(data = gbif3[gbif3$clade != "Birds", ], aes(x = decimallongitude, y = decimallatitude, 
                                  group = NA, color = clade), alpha = 0.5) +
  facet_wrap( ~ decade) +
  theme_void() +
  theme(legend.position = "top",
        legend.title = element_blank())
```

## Convex hulls

```{r, echo = FALSE}
# Create convex hulls for all species with >= 3 unique localities
hull_species <- list()
                              
for(i in 1:length(unique(gbif3$binomial))){
  hull_species[[i]] <- make.hull(gbif3[gbif3$binomial == unique(gbif3$binomial)[i], ])
}

# Add species names
names(hull_species) <- unique(gbif3$binomial)
```

Amphibians only, for all species with >= 3 unique localities (13 species)
```{r, echo = FALSE}
# Get amphibians species involved
gbif3 %>%
  filter(clade == "Amphibians") %>%
  pull(binomial) %>%
  unique()
```

```{r, echo = FALSE}
ggplot(data = world, aes(x = long, y = lat, group = group)) + 
  geom_polygon(fill = NA, col = "black", size = 0.5) +
  geom_polygon(data = hull_species[["Taricha granulosa"]], aes(x = x, y = y, group = NULL), 
              col = viridis(8)[13], fill = NA) +
  geom_polygon(data = hull_species[["Notophthalmus viridescens"]], aes(x = x, y = y, group = NULL),                  col = viridis(13)[1], fill = NA) +
  geom_polygon(data = hull_species[["Taricha torosa"]], aes(x = x, y = y, group = NULL),                             col = viridis(8)[2], fill = NA) +
  geom_polygon(data = hull_species[["Anaxyrus boreas"]], aes(x = x, y = y, group = NULL),                            col = viridis(8)[3], fill = NA) +
  geom_polygon(data = hull_species[["Pelophylax esculentus"]], aes(x = x, y = y, group = NULL),                      col = viridis(8)[4], fill = NA) +
  geom_polygon(data = hull_species[["Eurycea longicauda"]], aes(x = x, y = y, group = NULL),                      col = viridis(8)[5], fill = NA) +
  geom_polygon(data = hull_species[["Anaxyrus americanus"]], aes(x = x, y = y, group = NULL),                      col = viridis(8)[6], fill = NA) +
  geom_polygon(data = hull_species[["Aneides niger"]], aes(x = x, y = y, group = NULL),                      col = viridis(8)[7], fill = NA) +
  geom_polygon(data = hull_species[["Eurycea bislineata"]], aes(x = x, y = y, group = NULL),                      col = viridis(8)[8], fill = NA) +
  geom_polygon(data = hull_species[["Hyla arborea"]], aes(x = x, y = y, group = NULL),                      col = viridis(8)[9], fill = NA) +
  geom_polygon(data = hull_species[["Desmognathus fuscus"]], aes(x = x, y = y, group = NULL),                      col = viridis(8)[10], fill = NA) +
  geom_polygon(data = hull_species[["Pseudotriton ruber"]], aes(x = x, y = y, group = NULL),                      col = viridis(8)[11], fill = NA) +
  geom_polygon(data = hull_species[["Necturus maculosus"]], aes(x = x, y = y, group = NULL),                      col = viridis(8)[12], fill = NA) +
  theme_void() 
```
